Testing methods that raise an event
To test a method that raises an event, all you have to do is to subscribe to
that event in your test method before acting

// subscribe
'_logger.ErrorLogged += (sender, args) => { id = args; };'
// act
'_logger.Log("abc");'

In your event handler, you gonna have some value, in this case we have a
newGuid. So, if this event is raised, than our 'id' should no longer be an
empty Guid after acting.

And than follows a simple assertion that 'id' is not equal to an empty Guid.