In 'VideoService' class we extracted the code that is talking to an external resource into a separate class called 'FileReader'.
We also extracted an interface from this class called 'IFileReader'.

Than, we created another implementation of this interface called
'MockFileReader', and this is the one we are going to use in our unit tests.

And, finally, we decoupled 'VideoService' from 'FileReader()' by programming
against the interface.

There are three ways we can pass an instance of an class that implements
'IFileReader' interface in this 'VideoService':

1) To pass it as a parameter to the ReadVideoTitle() method

2) To pass it using a property

3) To pass it by a constructor parameter