Testing interaction between two objects
'PlaceOrder' method should store the order passed as the argument to the
'_storage'.

As we're injecting an 'IStorage' object in the constructor of the
'OrderService', in our test, we need to create a mock object for this
interface.
    var storage = new Mock<IStorage>();

And than we pass this 'storage.Object' as an argument to the
'OrderService' constructor.

Than we're acting:
    var order = new Order();
    service.PlaceOrder(order);

Than we need to assert that 'storage.Store' is called. We can use
'Setup' to program our mock, but we are using 'Verify' to check if the
right method is called with the right object.
    storage.Verify(s => s.Store(order));

Recap:
In order to test interaction between two objects, you can use the
'Verify' method of mock object.